+++ ruby/CHANGELOG.md
@@ -4,6 +4,7 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 
 ### Unreleased
 
+- feature: Adding `Alert` component
 - feature: Adding `Card` component
 - feature: Adding `Checkbox` component
 - feature: Adding `Input` component
@@ -11,6 +12,7 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 - feature: Adding `Separator` component
 - feature: Adding `Switch` component
 - fix: Removing `Phlex::Kit` from the base component
+- fix: Removing unecessary autoloading
 
 ### 0.3.1