+++ ruby/CHANGELOG.md
@@ -4,6 +4,10 @@ This is the changelog for the `essence` gem. For now, most of the updates will b
 
 ### Unreleased
 
+- TBA
+
+### 0.4.0
+
 - feature: Adding `Alert` component
 - feature: Adding `Checkbox` component
 - feature: Adding `Input` component